[MIG] l10n_it_delivery_note: Migration to 16.0
