+++ l10n_it_delivery_note/models/__init__.py
@@ -1,7 +1,10 @@
 from . import account_invoice
+from . import account_invoice_line
 from . import res_config_settings
 from . import res_partner
 from . import sale_order
+from . import sale_order_line
 from . import stock_delivery_note
+from . import stock_delivery_note_line
 from . import stock_location
 from . import stock_picking
