+++ l10n_it_delivery_note/security/res_users.xml
@@ -2,16 +2,14 @@
 <!--
     Copyright (c) 2020, Link IT Europe srl, Italy. All rights reserved.
   -->
-<odoo>
-    <data noupdate="True">
+<odoo noupdate="1">
 
-        <record id="base.user_root" model="res.users">
-            <field name="groups_id" eval="[(4, ref('can_change_number'))]" />
-        </record>
+    <record id="base.user_root" model="res.users">
+        <field name="groups_id" eval="[(4, ref('can_change_number'))]" />
+    </record>
 
-        <record id="base.user_admin" model="res.users">
-            <field name="groups_id" eval="[(4, ref('can_change_number'))]" />
-        </record>
+    <record id="base.user_admin" model="res.users">
+        <field name="groups_id" eval="[(4, ref('can_change_number'))]" />
+    </record>
 
-    </data>
 </odoo>
